@@ -1,6 +1,6 @@
 {
   "name": "@geovistory/design-system-web",
-  "version": "1.14.0",
+  "version": "1.14.1-pr-86.0",
   "description": "geovistory.org's design system: web components",
   "main": "dist/index.cjs.js",
   "module": "dist/index.js",
@@ -41,7 +41,7 @@
     "=========================== STENCIL ==========================": "",
     "generate": "npx stencil generate --config _dev.stencil.config.ts",
     "==================== BUILD ====================": "================================",
-    "build": "npm run prepare && npx stencil build --config _build.stencil.config.ts && rollup -c rollup.config.js",
+    "build": "npm run prepare && npx stencil build --config _build.stencil.config.ts && rollup -c rollup.config.mjs",
     "build:docs": "storybook build -o ../../storybook  --quiet",
     "build:stencil": "npm run prepare && npx stencil build --config _build.stencil.config.ts --docs",
     "build-watch:stencil": "npm run prepare && npx stencil build --config _build.stencil.config.ts --watch",
@@ -56,54 +56,54 @@
     "build-storybook": "storybook build"
   },
   "dependencies": {
+    "@rollup/plugin-typescript": "^11.1.2",
     "@stencil/core": "^4.0.1",
-    "patch-package": "^6.4.7",
+    "patch-package": "^7.0.2",
     "prismjs": "^1.29.0",
     "react-copy-to-clipboard": "^5.1.0",
     "react-icons": "^4.3.1",
     "react-syntax-highlighter": "^15.5.0",
-    "swiper": "^8.3.0"
+    "swiper": "^10.0.4"
   },
   "devDependencies": {
     "@babel/core": "^7.21.8",
     "@babel/preset-env": "^7.22.5",
     "@babel/preset-react": "^7.22.5",
     "@babel/preset-typescript": "^7.22.5",
-    "@ionic/core": "^6.7.5",
+    "@ionic/core": "^7.1.3",
     "@prettier/plugin-xml": "^3.0.0",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-node-resolve": "^13.1.3",
+    "@rollup/plugin-commonjs": "^25.0.3",
+    "@rollup/plugin-node-resolve": "^15.1.0",
     "@stencil/angular-output-target": "^0.7.1",
     "@stencil/react-output-target": "^0.5.3",
     "@stencil/sass": "^3.0.4",
-    "@storybook/addon-actions": "^7.0.24",
-    "@storybook/addon-essentials": "^7.0.24",
-    "@storybook/addon-interactions": "^7.0.24",
-    "@storybook/addon-links": "^7.0.24",
-    "@storybook/html-vite": "^7.0.24",
-    "@storybook/react": "^7.0.24",
-    "@storybook/react-webpack5": "^7.0.24",
+    "@storybook/addon-actions": "^7.1.0",
+    "@storybook/addon-essentials": "^7.1.0",
+    "@storybook/addon-interactions": "^7.1.0",
+    "@storybook/addon-links": "^7.1.0",
+    "@storybook/html-vite": "^7.1.0",
+    "@storybook/react": "^7.1.0",
+    "@storybook/react-webpack5": "^7.1.0",
     "@storybook/testing-library": "^0.2.0",
-    "@storybook/web-components-vite": "^7.0.24",
+    "@storybook/web-components-vite": "^7.1.0",
     "@types/jest": "^27.0.3",
-    "@types/plotly.js": "2.12.18",
+    "@types/plotly.js": "^2.12.24",
     "@types/prismjs": "^1.26.0",
-    "babel-loader": "^8.2.5",
-    "concurrently": "^7.2.2",
-    "ionicons": "^6.1.3",
-    "jest": "^27.4.5",
+    "babel-loader": "^9.1.3",
+    "concurrently": "^8.2.0",
+    "esbuild": "^0.18.14",
+    "ionicons": "^7.1.2",
+    "jest": "^27.0.3",
     "jest-cli": "^27.4.5",
-    "mermaid": "^10.2.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "rollup": "^2.79.0",
+    "mermaid": "^10.3.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "rollup": "^3.26.3",
     "rollup-plugin-copy": "^3.4.0",
     "rollup-plugin-multi-input": "^1.3.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-sourcemaps": "^0.6.3",
-    "rollup-plugin-typescript": "^1.0.1",
     "sass": "^1.63.6",
-    "storybook": "^7.0.24"
+    "storybook": "^7.1.0"
   },
   "license": "MIT",
   "gitHead": "8dc5228046d1120dd7dfad32f6891ff5ad126de3"
