@@ -1,7 +1,6 @@
 import commonjs from '@rollup/plugin-commonjs';
 import { nodeResolve } from '@rollup/plugin-node-resolve';
 import peerDepsExternal from 'rollup-plugin-peer-deps-external';
-import sourcemaps from 'rollup-plugin-sourcemaps';
 import multiInput from 'rollup-plugin-multi-input';
 import copy from 'rollup-plugin-copy';
 
@@ -18,11 +17,10 @@ export default {
     return id.startsWith('react') || id.startsWith('@geovistory');
   },
   plugins: [
-    multiInput({ relative: 'dist/components/' }),
+    multiInput.default({ relative: 'dist/components/' }),
     nodeResolve(),
     commonjs(),
     peerDepsExternal(),
-    sourcemaps(),
     copy({
       targets: [
         { src: 'dist/components/*.d.ts', dest: 'dist/components-cjs' },
