@@ -15,6 +15,6 @@ export default {
   },
 };
 
-export const ClassDistribution = stencilWrapper(
+export const ClassDistribution = await stencilWrapper(
   <geov-class-distri sparqlEndpoint="https://sparql.geovistory.org/api_v1_community_data" width={500} height={500}></geov-class-distri>,
 );
