@@ -13,9 +13,9 @@ export default {
     },
   },
 };
-export const EntityDownloadRdf = stencilWrapper(
+export const EntityDownloadRdf = await stencilWrapper(
   <geov-entity-download-rdf entityId="i785518" color="primary" expand="block" fill="outline" buttonLabel="Download RDF" buttonIcon="download-outline" />,
 );
-export const EntityDownloadRdf2 = stencilWrapper(
+export const EntityDownloadRdf2 = await stencilWrapper(
   <geov-entity-download-rdf entityId="i785518" color="warning" fill="outline" buttonLabel="Download" buttonIcon="download-sharp"></geov-entity-download-rdf>,
 );
